Function for grabbing a user's games
This lays all the groundwork for grabbing every game that a user is in. First, added a function that gets all entries from a collection (we pass in "Games" to get all the games). Then, a function to get entries in a collection that match the desired value (we pass in the player emails to find their games). Then a function for checking if the field in an individual doc is correct. Also had to add functions in the game viewer specifically for the games (like checking all 4 players).