Display Tables in the HTML Page -Logan
Had to rewrite the displayGames function so that it would run the functions in the correct order. Also made a makeTable function and a makeHeader function that together would create the table shown on the page. Also made a new div to host the table when you press the button